requirements.txt
crewai==0.28.8
langchain==0.1.20
langchain-core==0.1.49
langchain-community==0.0.38
langchain-google-genai==0.0.4
crewai-tools==0.1.6
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
Streamlit==1.49.1

 #pip install --upgrade crewai
# pip install 'crewai[tools]'
# python -m crew
# .\.venv\Scripts\Activate.ps1
#streamlit run app.py